Pre: split key types into separate module
