+++ src/store.rs
@@ -2,6 +2,7 @@ pub mod integer;
 pub mod integermulti;
 pub mod multi;
 pub mod single;
+pub mod keys;
 
 use lmdb::DatabaseFlags;
